pyproject.toml
[project]
name = "dpv-areas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.9.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "typer>=0.20.1",
]
